requirements.txt
asgiref==3.8.1
Django==5.0.10
mssql-django==1.5
pyodbc==5.2.0
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
python-dotenv~=1.0.1
cryptography~=44.0.0
reportlab~=4.2.5
matplotlib
json5~=0.10.0
requests~=2.32.3
bs4~=0.0.2
pywin32
beautifulsoup4~=4.12.3
utils~=1.0.2